Fix updater

+++ XwaManager/XwaManager/Updater.cs
@@ -3,6 +3,8 @@ using System.Diagnostics;
 using System.IO.Compression;
 using System.IO;
 using System.Net;
+using System.Windows;
+using System.Threading;
 
 namespace XwaManager;
 
@@ -40,47 +42,84 @@ internal static class Updater
         }
         catch
         {
+            Restart();
             return;
         }
 
-        File.Delete(bakFilePath);
-        File.Move(exeFilePath, bakFilePath);
-
-        var keepExtensions = new string[]
+        try
         {
-            ".zip",
-            ".bak",
-            ".json",
-        };
+            File.Delete(bakFilePath);
+            File.Move(exeFilePath, bakFilePath);
 
-        foreach (string file in Directory.EnumerateFiles(managerDirectory))
-        {
-            string fileName = Path.GetFileName(file);
-
-            bool found = false;
-
-            foreach (string ext in keepExtensions)
+            var keepExtensions = new string[]
             {
-                if (string.Equals(fileName, "XwaManager" + ext, StringComparison.OrdinalIgnoreCase))
+                ".zip",
+                ".bak",
+                ".json",
+            };
+
+            foreach (string file in Directory.EnumerateFiles(managerDirectory))
+            {
+                string fileName = Path.GetFileName(file);
+
+                bool found = false;
+
+                foreach (string ext in keepExtensions)
                 {
-                    found = true;
-                    break;
+                    if (string.Equals(fileName, "XwaManager" + ext, StringComparison.OrdinalIgnoreCase))
+                    {
+                        found = true;
+                        break;
+                    }
+                }
+
+                if (found)
+                {
+                    continue;
+                }
+
+                int retryCount = 10;
+
+                for (int i = 0; i <= retryCount; i++)
+                {
+                    try
+                    {
+                        File.Delete(file);
+                        break;
+                    }
+                    catch
+                    {
+                        if (i == retryCount)
+                        {
+                            throw;
+                        }
+
+                        Thread.Sleep(1000);
+                    }
                 }
             }
 
-            if (found)
-            {
-                continue;
-            }
+            ZipFile.ExtractToDirectory(zipFilePath, managerDirectory);
+        }
+        catch (Exception ex)
+        {
+            MessageBox.Show(ex.ToString());
 
-            File.Delete(file);
+            if (File.Exists(bakFilePath))
+            {
+                File.Delete(bakFilePath);
+            }
         }
 
-        ZipFile.ExtractToDirectory(zipFilePath, managerDirectory);
+        Restart();
+    }
 
+    public static void RestartAutoUpdate()
+    {
         using (var process = Process.GetCurrentProcess())
         {
-            Process.Start(process.MainModule.FileName, process.StartInfo.Arguments);
+            string arguments = process.StartInfo.Arguments + " autoupdate";
+            Process.Start(process.MainModule.FileName, arguments);
         }
 
         Environment.Exit(0);
@@ -88,9 +127,18 @@ internal static class Updater
 
     public static void Restart()
     {
+        MessageBox.Show("Restart()");
+
         using (var process = Process.GetCurrentProcess())
         {
-            Process.Start(process.MainModule.FileName, process.StartInfo.Arguments);
+            string arguments = process.StartInfo.Arguments;
+
+            if (arguments.EndsWith(" autoupdate", StringComparison.OrdinalIgnoreCase))
+            {
+                arguments = arguments[..^" autoupdate".Length];
+            }
+
+            Process.Start(process.MainModule.FileName, arguments);
         }
 
         Environment.Exit(0);
